shared/requirements.txt
PyJWT
python-dotenv
